+++ test/profile/test_profile.py
@@ -1,12 +1,22 @@
+import os.path
+
 import torch
 import torch.nn.functional as F
+from torch.profiler import ProfilerActivity, profile
 
 from torch_geometric.nn import GraphSAGE
-from torch_geometric.profile import get_stats_summary, profileit, timeit
-from torch_geometric.testing import withCUDA
+from torch_geometric.profile import (
+    get_stats_summary,
+    profileit,
+    rename_profile_file,
+    timeit,
+    trace_handler,
+)
+from torch_geometric.testing import onlyFullTest, withCUDA
 
 
 @withCUDA
+@onlyFullTest
 def test_profile(get_dataset):
     dataset = get_dataset(name='PubMed')
     data = dataset[0].cuda()
@@ -56,3 +66,21 @@ def test_profile(get_dataset):
     assert stats_summary.max_active_cuda > 0
     assert stats_summary.min_nvidia_smi_free_cuda > 0
     assert stats_summary.max_nvidia_smi_used_cuda > 0
+
+
+@onlyFullTest
+def test_trace_handler(get_dataset):
+    dataset = get_dataset(name='PubMed')
+    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
+    data = dataset[0].to(device)
+    model = GraphSAGE(dataset.num_features, hidden_channels=64, num_layers=3,
+                      out_channels=dataset.num_classes).to(device)
+    model.eval()
+
+    for epoch in range(3):
+        with profile(activities=[ProfilerActivity.CPU, ProfilerActivity.CUDA],
+                     on_trace_ready=trace_handler) as p:
+            model(data.x, data.edge_index)
+            p.step()
+    rename_profile_file('test_profile')
+    assert os.path.exists('profile-test_profile.json')